yask: Refactor entire backend
Cleaner, more structured, more logically correct,
less redundant code, better names, more pythonic
